Move Usage indexing and constants out of webhook package
This way the controller package doesn't need to import the webhook
package.
